+++ Oprim.Domain/Entities/PMO/ProjectItem.cs
@@ -9,10 +9,10 @@ public class ProjectItem : BaseProjectEntity
     
 
     [ForeignKey("ProjectItemGroupId")] public ProjectItemGroup ProjectItemGroup { get; set; }
-    public int ProjectItemGroupId { get; set; }
+    public long ProjectItemGroupId { get; set; }
 
     [ForeignKey("ProjectCostBreakdownId")] public ProjectCostBreakdown ProjectCostBreakdown { get; set; }
-    public int ProjectCostBreakdownId { get; set; }
+    public long ProjectCostBreakdownId { get; set; }
 
     public string Code { get; set; }
